[6.5]: Tích hợp RabbitMQ vào Ordering.API Microservices – Consumer

+++ src/Services/Ordering/Ordering.Application/Features/V1/Orders/Commands/DeleteOrder/DeleteOrderCommandHandle.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
 using MediatR;
+using Ordering.Application.Common.Exceptions;
 using Ordering.Application.Common.Interfaces;
+using Ordering.Domain.Entities;
 using ILogger = Serilog.ILogger;
 
 namespace Ordering.Application.Features.V1.Orders.Commands.DeleteOrder
@@ -21,6 +23,9 @@ namespace Ordering.Application.Features.V1.Orders.Commands.DeleteOrder
 
         public async Task<Unit> Handle(DeleteOrderCommand request, CancellationToken cancellationToken)
         {
+            var orderEntity = await _oderRepository.GetByIdAsync(request.Id);
+            if (orderEntity is null) throw new NotFoundException(nameof(Order), request.Id);
+
             _logger.Information($"BEGIN: {MethodName}");
 
             await _oderRepository.DeleteOrder(request.Id);
